% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.delimit.R
\name{comp.delimit}
\alias{comp.delimit}
\title{Title comp.delimit}
\usage{
comp.delimit(obj, alt, method = "EM")
}
\arguments{
\item{obj}{}

\item{alt}{}

\item{method}{}
}
\value{

}
\description{
Title comp.delimit
}
