% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmyc.R
\name{gmyc}
\alias{gmyc}
\title{Titlegmyc}
\usage{
gmyc(tr, method = "single", interval = c(0, 5), quiet = FALSE)
}
\arguments{
\item{tr}{}

\item{method}{}

\item{interval}{}

\item{quiet}{}
}
\value{

}
\description{
Titlegmyc
}
