% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_nupgen-package.R
\docType{package}
\name{splitsNupgen-package}
\alias{splitsNupgen}
\alias{splitsNupgen-package}
\title{splitsNupgen: SPecies' LImits by Threshold Statistics}
\description{
"splits" contains tools for delimiting species and automated taxonomy at multiple levels of biological organization (eg. DNA barcoding, morphometrics), both top-down (merging phylogenetic and phylogeographic methods) and bottom-up (single samples into >1 groups).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/splits_nupgen}
  \item Report bugs at \url{https://github.com/brunomioto/splits_nupgen/issues}
}

}
\keyword{internal}
