% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gmyc.R
\name{summary.gmyc}
\alias{summary.gmyc}
\title{Title summary.gmyc}
\usage{
\method{summary}{gmyc}(object, second.peak = FALSE, ...)
}
\arguments{
\item{object}{}

\item{second.peak}{}

\item{...}{}
}
\value{

}
\description{
Title summary.gmyc
}
